# contracts/openapi/core-api.v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "kinsun.ai Core API",
    "version": "1.1.0",
    "summary": "Implemented Core Domain, consent, security and outbox APIs.",
    "description": "Current executable Core API contract. Every protected operation re-evaluates tenant, elder, relationship or assignment scope. Cognito JWT verification is not yet wired; production therefore fails closed with 401 as documented by bearerAuth."
  },
  "paths": {
...
    "/api/v1/internal/elders/{elder_id}/summary-drafts": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Create Summary Draft",
        "operationId": "create_summary_draft_api_v1_internal_elders__elder_id__summary_drafts_post",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Elder Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/domain/CreateSummaryDraftRequestV1.json"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/DailySummaryEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/elders/{elder_id}/summaries": {
      "get": {
        "tags": [
          "summaries"
        ],
        "summary": "List Summaries",
        "description": "List formal summaries by default; expose review states only to reviewers.",
        "operationId": "list_summaries_api_v1_elders__elder_id__summaries_get",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Elder Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "DRAFT",
                      "READY",
                      "NEEDS_REVIEW",
                      "PUBLISHED",
                      "STALE",
                      "WITHDRAWN"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to formal READY and PUBLISHED summaries. Any explicit non-formal status additionally requires summary:review.",
              "title": "Status"
            },
            "description": "Defaults to formal READY and PUBLISHED summaries. Any explicit non-formal status additionally requires summary:review."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/DailySummaryListEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/elders/{elder_id}/summaries/{summary_id}": {
      "get": {
        "tags": [
          "summaries"
        ],
        "summary": "Get Summary",
        "description": "Return a formal summary, or a non-formal summary only to a reviewer.",
        "operationId": "get_summary_api_v1_elders__elder_id__summaries__summary_id__get",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Elder Id"
            }
          },
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/DailySummaryEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/elders/{elder_id}/summaries/{summary_id}/review": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Review Summary",
        "operationId": "review_summary_api_v1_elders__elder_id__summaries__summary_id__review_post",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Elder Id"
            }
          },
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/domain/ReviewSummaryRequestV1.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/SummaryReviewEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/elders/{elder_id}/summaries/{summary_id}/rebuild": {
      "post": {
        "tags": [
          "summaries"
        ],
        "summary": "Rebuild Summary",
        "operationId": "rebuild_summary_api_v1_elders__elder_id__summaries__summary_id__rebuild_post",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Elder Id"
            }
          },
          {
            "name": "summary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Summary Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/domain/RebuildSummaryRequestV1.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/DailySummaryEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Target JWT shape. A production verifier is not configured yet; protected routes fail closed with 401. Development fake auth requires the explicit FAKE_AUTH_ENABLED flag."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or authenticator unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The actor is inactive or its role cannot perform the operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "NotFoundOrForbidden": {
        "description": "Resource missing or outside live authorization scope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Conflict": {
        "description": "State, version, or idempotency conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Schema or semantic validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Required dependency unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}